requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
python-gnupg
azure.identity
azure.keyvault.secrets
azure.storage.blob
smart_open